# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GetCurrentAuthUserResponseUser = {
  id?: string;
  name?: string;
  profileImage?: string;
} | null;

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: GetCurrentAuthUserResponseUser;
}

export type UserProfileNaicsCodesItem = {
  code: string;
  title: string;
  relevance: string;
};

export interface UserProfile {
  id: string;
  businessName?: string;
  onboarded: boolean;
  entityType?: string;
  state?: string;
  industry?: string;
  stage?: string;
  ownerName?: string;
  contactEmail?: string;
  contactPhone?: string;
  zipCode?: string;
  yearsInBusiness?: string;
  employees?: string;
  annualRevenue?: string;
  fundingGoals?: string[];
  missionStatement?: string;
  certifications?: string[];
  fundingAmount?: string;
  demographics?: string[];
  naicsCodes?: UserProfileNaicsCodesItem[];
}

export type UpdateProfileBodyNaicsCodesItem = {
  code: string;
  title: string;
  relevance: string;
};

export interface UpdateProfileBody {
  businessName?: string;
  onboarded?: boolean;
  entityType?: string;
  state?: string;
  industry?: string;
  ownerName?: string;
  contactEmail?: string;
  contactPhone?: string;
  zipCode?: string;
  yearsInBusiness?: string;
  employees?: string;
  annualRevenue?: string;
  fundingGoals?: string[];
  missionStatement?: string;
  certifications?: string[];
  fundingAmount?: string;
  naicsCodes?: UpdateProfileBodyNaicsCodesItem[];
}

export interface ChecklistState {
  [key: string]: boolean;
}

export interface UpdateChecklistBody {
  itemId: string;
  completed: boolean;
}

export type AiChatBodyHistoryItem = {
  role: string;
  content: string;
};

export interface AiChatBody {
  message: string;
  history?: AiChatBodyHistoryItem[];
}

export interface AiChatResponse {
  reply: string;
}

export interface AiNaicsBody {
  query: string;
  industry?: string;
}

export type AiNaicsResponseCodesItem = {
  code: string;
  title: string;
  description: string;
  relevance: string;
  govContractTip: string;
};

export interface AiNaicsResponse {
  codes: AiNaicsResponseCodesItem[];
}

export type AiGrantsBodyProfile = { [key: string]: unknown };

export interface AiGrantsBody {
  profile: AiGrantsBodyProfile;
}

export type AiGrantsResponseGrantsItem = { [key: string]: unknown };

export interface AiGrantsResponse {
  grants: AiGrantsResponseGrantsItem[];
}

export type LoginParams = {
  returnTo?: string;
};