# lib/api-zod/src/generated/types/aiGrantsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiGrantsBodyProfile } from "./aiGrantsBodyProfile";

export interface AiGrantsBody {
  profile: AiGrantsBodyProfile;
}