# lib/api-zod/src/generated/types/aiGrantsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiGrantsResponseGrantsItem } from "./aiGrantsResponseGrantsItem";

export interface AiGrantsResponse {
  grants: AiGrantsResponseGrantsItem[];
}