# lib/api-zod/src/generated/types/aiNaicsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AiNaicsBody {
  query: string;
  industry?: string;
}