# lib/api-zod/src/generated/types/getCurrentAuthUserResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GetCurrentAuthUserResponseUser } from "./getCurrentAuthUserResponseUser";

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: GetCurrentAuthUserResponseUser;
}