# lib/api-zod/src/generated/types/updateChecklistBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateChecklistBody {
  itemId: string;
  completed: boolean;
}