# lib/api-zod/src/generated/types/updateProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BizLaunch API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProfileBodyNaicsCodesItem } from "./updateProfileBodyNaicsCodesItem";

export interface UpdateProfileBody {
  businessName?: string;
  onboarded?: boolean;
  entityType?: string;
  state?: string;
  industry?: string;
  ownerName?: string;
  contactEmail?: string;
  contactPhone?: string;
  zipCode?: string;
  yearsInBusiness?: string;
  employees?: string;
  annualRevenue?: string;
  fundingGoals?: string[];
  missionStatement?: string;
  certifications?: string[];
  fundingAmount?: string;
  naicsCodes?: UpdateProfileBodyNaicsCodesItem[];
}